Use BlockDTO
