--- lib/src/main/java/com/thenewboston/banktransactions/model/BankBlock.kt
@@ -1,20 +1,0 @@
-package com.thenewboston.banktransactions.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class BankBlock(
-    @SerialName("id")
-    val id: String,
-    @SerialName("balance_key")
-    val balanceKey: String,
-    @SerialName("sender")
-    val sender: String,
-    @SerialName("signature")
-    val signature: String,
-    @SerialName("created_date")
-    val created: String,
-    @SerialName("modified_date")
-    val modified: String
-)
